+++ MOP/src/FSM/FsmManager.cs
@@ -18,7 +18,6 @@ using HutongGames.PlayMaker;
 using System.Reflection;
 using UnityEngine;
 
-using MOP.FSM;
 using MOP.Common.Enumerations;
 using MOP.Managers;