+++ MOP/src/Helpers/SaveManager.cs
@@ -219,7 +219,7 @@ namespace MOP.Helpers
 
             if (saveBugs.Count > 0)
             {
-                ModUI.ShowYesNoMessage($"MOP found <color=yellow>{saveBugs.Count}</color> problem{(saveBugs.Count > 1 ? "s" : "")} with your save:\n\n" +
+                ModUI.ShowYesNoMessage($"MOP has found <color=yellow>{saveBugs.Count}</color> problem{(saveBugs.Count > 1 ? "s" : "")} with your save:\n\n" +
                                        $"<color=yellow>{string.Join(", ", saveBugs.Select(f => f.BugName).ToArray())}</color>\n\n" +
                                        $"Would you like MOP to try and fix {((saveBugs.Count > 1) ? "them" : "it")}?", "MOP - Save Integrity Verification", FixAllProblems);
             }